chore: fixing some pylint errors
