+++ reptar/calculators/psi4_workers.py
@@ -25,7 +25,7 @@ import os
 from collections.abc import Iterable
 from tempfile import TemporaryDirectory
 import numpy as np
-from . import Data
+from .data import Data
 from .utils import initialize_worker_data
 from ..logger import ReptarLogger
 from ..parsers.gaussian_cube import parse_cube
