+++ reptar/calculators/xtb_workers.py
@@ -26,7 +26,7 @@ import os
 import subprocess
 from tempfile import TemporaryDirectory
 import numpy as np
-from . import Data
+from .data import Data
 from .utils import initialize_worker_data
 from ..writers.xyz import write_xyz
 from ..utils import parse_xyz
